Fix bug where background updates not reflected
Still not exactly sure what's causing this problem, something about the
connection pool keeping connections alive, and having some implicit
transaction open, which causes updates input manually to the sqlite3 cli
to not be shown. Adding an idle_timeout to the connection pool fixes the
issue, but also defeats part of the purpose of using a connection pool,
so this should probably be re-visited.